harness-resizer/requirements.txt
PyGithub>=1.55
aiohttp>=3.8.1
click>=8.1.3
dataclasses-json>=0.5.7
oyaml>=1.0
prettytable>=3.3.0